11 - resolve PR returns commentaires
