+++ babel.config.json
@@ -1,6 +1,7 @@
 {
     "presets": [
-        "@babel/react",
-        "@babel/env"
-    ]
+        "@babel/preset-env",
+        "@babel/preset-react"
+    ],
+    "plugins": ["@babel/proposal-class-properties"]
 }
