+++ package.json
@@ -28,12 +28,13 @@
   },
   "homepage": "https://github.com/reactjs-academy/redux-saga-exercise",
   "devDependencies": {
+    "@babel/core": "^7.0.0",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-transform-runtime": "^7.19.6",
-    "babel-cli": "^6.16.0",
-    "babel-core": "^6.0.20",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^4.1.3",
-    "babel-loader": "^6.2.8",
+    "babel-jest": "^29.3.1",
+    "babel-loader": "^9.1.0",
     "babel-plugin-syntax-flow": "^6.18.0",
     "babel-plugin-transform-flow-strip-types": "^6.18.0",
     "babel-polyfill": "^6.16.0",
@@ -55,6 +56,7 @@
     "react-hot-loader": "^1.3.1",
     "redux-logger": "^2.6.1",
     "redux-promise": "^0.5.3",
+    "regenerator-runtime": "^0.13.11",
     "style-loader": "^0.13.1",
     "webpack": "^1.13.3",
     "webpack-dev-middleware": "^1.8.4",
@@ -62,12 +64,17 @@
     "webpack-hot-middleware": "^2.13.0"
   },
   "dependencies": {
-    "@babel/core": "^7.20.5",
+    "@babel/cli": "^7.19.3",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-es2015": "^7.0.0-beta.53",
     "@babel/preset-react": "^7.18.6",
     "@testing-library/dom": "^8.19.0",
     "@testing-library/react": "12.0.0",
+    "babel-cli": "^6.26.0",
+    "babel-plugin-lodash": "^3.3.4",
+    "babel-plugin-react-transform": "^3.0.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-upgrade": "^1.0.1",
     "es6-promise": "^4.0.5",
     "eslint": "^7.32.0 || ^8.2.0",
     "eslint-config-airbnb": "^19.0.4",
