+++ src/components/Navigations/Menus/__tests__/BeersLiked.spec.js
@@ -2,27 +2,31 @@ import React from "react";
 
 import { expect } from "chai";
 
-import { storeEmptyLikedBeerInit, propsLikedBeerInit } from "../../../../__tests__/initValues";
 import BeersLiked from "../Beersliked";
-import { queryAllByTestId, render} from "../../../../__tests__/customeMethodRTL";
+import { initListBeers, propsLikedBeerEmpty, propsLikedBeersNotEmpty } from "../../../../utils/testTools/initValues";
+import { queryAllByTestId, render} from "../../../../utils/testTools/customeMethodRTL";
 
 describe("BeersLiked component should display all liked beer title", () => {
-    it("Should display only a div if liked beers array is empty", () => {
+  it("Should display only a div if liked beers array is empty", () => {
 
-        render (
-            <BeersLiked />,
-            storeEmptyLikedBeerInit
-        );
+    render (
+      <BeersLiked
+        likes={ propsLikedBeerEmpty }
+        beers={ initListBeers }
+      />
+    );
 
-        expect(queryAllByTestId("beer-title").length).to.be.equal(0);
-    });
+    expect(queryAllByTestId("beer-title").length).to.be.equal(0);
+  });
 
-    it("Should return same number of beer-title than number of beers liked", () => {
-        render(
-            <BeersLiked/>,
-            propsLikedBeerInit
-        );
+  it("Should return same number of beer-title than number of beers liked", () => {
+    render(
+      <BeersLiked
+        likes={ propsLikedBeersNotEmpty }
+        beers={ initListBeers }
+      />
+    );
 
-        expect(queryAllByTestId("beer-title").length).to.be.equal(13);
-    })
+    expect(queryAllByTestId("beer-title").length).to.be.equal(13);
+  })
 });
