+++ src/config/index.js
@@ -1,3 +1,3 @@
 // @flow
-export const API_BASE_URL = 'http://localhost:3000'
-export const BEER_LIMIT_PER_PAGE = 10
+export const API_BASE_URL = 'http://localhost:3000';
+export const BEER_LIMIT_PER_PAGE = 10;
