+++ src/sagas/addBeerLikeSaga.js
@@ -4,10 +4,10 @@ import { ADD_BEER_LIKE, setCurrentBeerIndex } from "../actions";
 import { getCurrentIndex } from "../selectors";
 
 export default function* addLikesSaga (action) {
-    yield put({ type: ADD_BEER_LIKE, beerId: action.beerId});
+  yield put({ type: ADD_BEER_LIKE, beerId: action.beerId});
 
-    // next()
-    const currentIndex = yield select(getCurrentIndex);
-    yield put(setCurrentBeerIndex(currentIndex + 1));
+  // next()
+  const currentIndex = yield select(getCurrentIndex);
+  yield put(setCurrentBeerIndex(currentIndex + 1));
 };
