+++ src/sagas/index.js
@@ -1,12 +1,12 @@
-import { takeLeading, takeEvery } from 'redux-saga/effects';
+import { takeLeading } from 'redux-saga/effects';
 
-import { START_SESSION, FETCH_BEERS, ADD_LIKE } from '../actions';
-import fetchStartSession from './sagaSession';
-import * as rootBeers from "./sagaBeers";
+import fetchStartSession from './startSessionSaga';
 import addLikesSaga from "./addBeerLikeSaga";
+import * as rootBeers from "./fetchBeersSaga";
+import { START_SESSION, FETCH_BEERS, ADD_LIKE } from '../actions';
 
 export default function* rootSaga() {
-    yield takeLeading(START_SESSION, fetchStartSession)
-    yield takeLeading(FETCH_BEERS, rootBeers.fetchBeersSaga)
-    yield takeLeading(ADD_LIKE, addLikesSaga)
+  yield takeLeading(START_SESSION, fetchStartSession)
+  yield takeLeading(FETCH_BEERS, rootBeers.fetchBeersSaga)
+  yield takeLeading(ADD_LIKE, addLikesSaga)
 };
