+++ src/selectors/index.js
@@ -1,11 +1,11 @@
 export const isFetchingBeers = (state) => (
   state.beers.isFetching
-)
+);
 
 export const settings = (state) => (
   state.settings
-)
+);
 
 export const getCurrentIndex = (state) => (
   state.beer.currentIndex
-)
+);
