--- src/utils/__tests__/renderMethodRTL.js
@@ -1,5 +1,0 @@
-import { screen } from "@testing-library/react";
-
-export const getByClass = (className) => {
-    return screen.getByClass(className);
-};
