+++ src/utils/getInfoBeer.js
@@ -1,4 +1,3 @@
 export const getBeerInfoByIndex = (beerId, listBeers) => {
-    const result = listBeers.find(beer => beer.id == beerId);
-    return result
+  return listBeers.find(beer => beer.id === beerId);
 };
